QtAviSynth: Improve the image quality of zoomnpan/perspective
The quat2quad transform implementations which I tried out have in
common that the image quality deteriorates the smaller the target
rect is compared to the source quad. No matter which interpolater
is being used.

A hack to improve it is to make the targte rect fill the source
video size and then resample the transform result to the target
size. I am not happy about double resampling, but the result is OK.
As resampling algorithm, I found that "Blackman" with 1 tap gave
the best results for my use case.