QmlAnimationRenderer: Simplify qml clip duration identification
Using different possible sources for determining the animation
duration (and as consequence of the frame count) was buggy and
confused my just a few months after hacking that.

Let's define hereby that the timeline length is interpreted as
clip length in milliseconds and ignore what a possible timeline
animation might say otherwise.

Milliseconds is anyways the default that is set up by Qt Design
Studio.